refactor(#417): extract the home hero and cover its session branches
